Moving tests/unit/states/test_ldap.py to pytest.
